Changed User class to struct
