+++ Src/BootCamp.Chapter/User.cs
@@ -1,14 +1,13 @@
 ﻿namespace BootCamp.Chapter
 {
-    internal class User
+    internal struct User
     {
-        public string UserName { get; private set; }
-
-        public string Password { get; private set; }
+        public readonly string Username;
+        public readonly string Password;
 
         public User(string username, string password)
         {
-            UserName = username;
+            Username = username;
             Password = password;
         }
     }
